Move DistributedClosestPoint definitions into a .cpp file.
This is to avoid unneeded parsing of the code.
